% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blobs}
\alias{blobs}
\title{8 blob like structures}
\format{A data frame with 10 rows and 800 variables}
\usage{
blobs
}
\description{
A simulated dataset of 8 Gaussian blobs. Simulated using 
the 'clusterlab' CRAN package.
}
\keyword{datasets}
