% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brain}
\alias{brain}
\title{A brain cancer dataset}
\format{A list of data frames}
\source{
\url{https://gdac.broadinstitute.org/}
}
\usage{
brain
}
\description{
A dataset containing The Cancer Genome Atlas expression
data. From this publication https://tcga-data.nci.nih.gov/docs/publications/lgggbm_2016/.
}
\keyword{datasets}
