% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo.R
\name{build_gabriel_graph}
\alias{build_gabriel_graph}
\title{constructs a Gabriel graph based on spatial location information}
\usage{
build_gabriel_graph(
  position,
  data_struct = "dist",
  dist.th = 2,
  deldir.wt = F,
  deldir.nb = 3
)
}
\arguments{
\item{position}{A matrix or data frame of coordinates, with each row representing
a point and two columns for x and y coordinates.}

\item{data_struct}{A character string specifying the structure to use.
Options are "deldir" for Delaunay triangulation
or "dist" (default) for distance threshold.}

\item{dist.th}{A numeric threshold for the distance used when data_struct is set
to "dist". Defaults to 2.}

\item{deldir.wt}{Logical. If TRUE, distances between neighboring points are returned
as weights in the Gabriel graph. Defaults to FALSE, which returns
a logical graph.}

\item{deldir.nb}{An integer specifying the threshold for the number of neighbors to
consider in the Gabriel graph construction. Defaults to 3.}
}
\value{
A sparse adjacency matrix representing the Gabriel graph, where rows and
columns correspond to nodes and non-zero entries indicate edges between nodes.
}
\description{
constructs a Gabriel graph based on spatial location information
}
