% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo.R
\name{cell_groups}
\alias{cell_groups}
\title{Identify Cell Groups Based on Gene Expression and Spatial Connectivity}
\usage{
cell_groups(st_data, maxVol = 0.975, final_GG, thres)
}
\arguments{
\item{st_data}{A matrix or data frame containing spatial transcriptomics data
with genes as rows and spatial spots as columns.}

\item{maxVol}{Removes genes present in this fraction or more of spots}

\item{final_GG}{A sparse adjacency matrix representing the Gabriel Graph of
spatial spots, where edges indicate neighboring spots.}

\item{thres}{A list of threshold values for each gene, used to identify cells or
spots with significant gene expression.}
}
\value{
A data frame with two main metrics for each gene meeting the volume threshold:
- "Volume ratio": Proportion of spatial spots with expression above a
threshold for each gene.
- "Avg #Neighbors ratio": Average neighbor count ratio for each gene,
reflecting the spatial clustering of expressing cells.
}
\description{
Identify Cell Groups Based on Gene Expression and Spatial Connectivity
}
