\name{celltype_predict}
\alias{celltype_predict}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
predict cell type
}
\description{
predicted cell type
}
\usage{
celltype_predict(matrix)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{matrix}{
matrix Rows should be cells and columns should be genes.
}
}

\value{
return predicted cell type
}

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x)
{
  }
}

