% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo.R
\name{check_graph}
\alias{check_graph}
\title{Cenerates and plots a Gabriel Graph based on spatial data, using Delaunay triangulation or a distance threshold.}
\usage{
check_graph(
  position,
  data_struct = "dist",
  dist.th = 2,
  deldir.nb = 3,
  vertex.size = 1,
  ratio = 1
)
}
\arguments{
\item{position}{A matrix or data frame of coordinates, with each row representing
a point and two columns for x and y coordinates.}

\item{data_struct}{A character string specifying the structure to use.
Options are "deldir" for Delaunay triangulation
or "dist" (default) for distance threshold.}

\item{dist.th}{A numeric threshold for the distance used when data_struct is set
to "dist". Defaults to 2.}

\item{deldir.nb}{An integer specifying the threshold for the number of neighbors to
consider in the Gabriel graph construction. Defaults to 3.}

\item{vertex.size}{Numeric. Controls the size of the vertices in the plot. Default is 1.}

\item{ratio}{Numeric. Aspect ratio for the plot. Default is 1.}
}
\value{
A plot of the Gabriel Graph showing the spatial connections between spots
}
\description{
Cenerates and plots a Gabriel Graph based on spatial data, using Delaunay triangulation or a distance threshold.
}
