% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{circles}
\alias{circles}
\title{Three concentric circles}
\format{A data frame with 2 rows and 540 variables}
\usage{
circles
}
\description{
Simulated data using the 'clusterSim' CRAN package.
}
\keyword{datasets}
