% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_functions.R
\name{CNN_kernel_mine_b}
\alias{CNN_kernel_mine_b}
\title{CNN_kernel_mine_b: fast adaptive density aware kernel}
\usage{
CNN_kernel_mine_b(mat, NN = 3, NN2 = 7)
}
\arguments{
\item{mat}{Matrix: matrix should have samples as columns and rows as features}

\item{NN}{Numerical value: the number of nearest neighbours to use when calculating local sigma}

\item{NN2}{Numerical value: the number of nearest neighbours to use when calculating common nearest neighbours}
}
\value{
A kernel matrix
}
\description{
CNN_kernel_mine_b: fast adaptive density aware kernel
}
\examples{
CNN_kern <- CNN_kernel_mine_b(blobs[,1:50])
}
