% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{get_cp_lr}
\alias{get_cp_lr}
\title{Calculate ligand-receptor associations with cell-type Pairs in colocalized spots}
\usage{
get_cp_lr(lr_sp_df, pred_prop)
}
\arguments{
\item{lr_sp_df}{A matrix containing ligand-receptor expression data, with spots as rows and ligand-receptor pairs as columns.}

\item{pred_prop}{A matrix containing cell type proportions, with rows representing
spots and columns representing cell types.}
}
\value{
A data frame summarizing significant ligand-receptor and cell-type pair
associations, including p-values, number of colocalized spots, geometric mean expression,
and an indication of autocrine signaling.
}
\description{
Calculate ligand-receptor associations with cell-type Pairs in colocalized spots
}
