% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo.R
\name{get_deconvo_matrix}
\alias{get_deconvo_matrix}
\title{Performs deconvolution on a given input feature matrix and weight matrix of ST data}
\usage{
get_deconvo_matrix(Xinput, K, method = "brunet")
}
\arguments{
\item{Xinput}{A filtered and processed feature matrix with rows representing markers
and columns representing spots.}

\item{K}{A weight matrix with rows representing markers and columns
representing cell types.}

\item{method}{The NMF algorithm to use, with "brunet" as the default.}
}
\value{
A matrix representing the proportions of each cell type in each spatial
spot. Rows represent spatial spots and columns represent cell types.
}
\description{
Performs deconvolution on a given input feature matrix and weight matrix of ST data
}
