% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo.R
\name{get_feature_matrix}
\alias{get_feature_matrix}
\title{Creates a feature matrix from ST data by selecting markers relevant to specific cell types as features}
\usage{
get_feature_matrix(st_data, markerList)
}
\arguments{
\item{st_data}{A matrix or data frame containing spatial transcriptomics data
with genes as rows and spatial spots as columns.}

\item{markerList}{A list where each element is a vector of markers
for a specific cell type.}
}
\value{
A filtered and processed feature matrix with rows representing markers
and columns representing spots.
}
\description{
Creates a feature matrix from ST data by selecting markers relevant to specific cell types as features
}
