% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo.R
\name{get_filter}
\alias{get_filter}
\title{Filter proportion matrix based on threshold}
\usage{
get_filter(P, filt = 0.05)
}
\arguments{
\item{P}{A matrix representing the proportion of each cell type at each spot.
Rows represent spatial spots and columns represent cell types.}

\item{filt}{A numeric value representing the filtering threshold. Values in
proportion matrix that are less than or equal to this fraction are
set to zero. Default is 0.05.}
}
\value{
a filtered proportion matrix
}
\description{
Filter proportion matrix based on threshold
}
