% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_position}
\alias{get_position}
\title{Convert Position IDs to Coordinate Data Frame}
\usage{
get_position(pos_id, sep = "x")
}
\arguments{
\item{pos_id}{A character vector of position IDs, each representing coordinates
separated by a delimiter (e.g., "123x456").}

\item{sep}{A string delimiter used to separate x and y coordinates in each position ID.
Default is "x".}
}
\value{
A data frame with numeric columns "x" and "y" representing coordinates.
Row names correspond to the original position ID strings.
}
\description{
Convert Position IDs to Coordinate Data Frame
}
\examples{
get_position(c("10x20", "30x40"))
get_position(c("5_8", "12_16"), sep = "_")

}
