% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{get_spatial_lr}
\alias{get_spatial_lr}
\title{Computes strength of ligand-receptor interaction for each spatial spot.}
\usage{
get_spatial_lr(st_data, spots, cc_lr, database)
}
\arguments{
\item{st_data}{A matrix or data frame containing spatial transcriptomics data
with genes as rows and spatial spots as columns.}

\item{spots}{A character vector of selected spot names from the spatial data.}

\item{cc_lr}{A data frame containing ligand-receptor pairs with columns named
"ligand", "receptor", and "LR" (ligand-receptor pair names).}

\item{database}{Prior knowledge of signaling pathways (Omnipath_interactions)}
}
\value{
A matrix of ligand-receptor interaction scores for each spot.
}
\description{
Computes strength of ligand-receptor interaction for each spatial spot.
}
