% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo.R
\name{get_weight_matrix}
\alias{get_weight_matrix}
\title{Generate Weight Matrix for features}
\usage{
get_weight_matrix(markerList, score)
}
\arguments{
\item{markerList}{A list where each element is a vector of markers
for a specific cell type.}

\item{score}{A named vector of scores, where the names correspond to genes and
the values are associated scores.}
}
\value{
A normalized weight matrix with rows representing markers and columns
representing cell types.
}
\description{
Generate Weight Matrix for features
}
