% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{harmonise_ids}
\alias{harmonise_ids}
\title{harmonise_ids: works on a list of similarity matrices to add entries of NA where
there are missing observations between views}
\usage{
harmonise_ids(l)
}
\arguments{
\item{l}{A list of similarity matrices: those to be harmonised.}
}
\value{
A list of harmonised similarity matrices.
}
\description{
Simply adds a column and row of NA with the missing ID for data imputation. The
similarity matrix requires row and column IDs present for this to work.
}
\examples{
h_test <- harmonise_ids(missl) 
}
