% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{integrate}
\alias{integrate}
\title{Integrate spatial neighbors based on distance or rank}
\usage{
integrate(
  position,
  mode = "dist",
  rank.n = 2,
  th.min = 0,
  th.max = 1,
  plot = T
)
}
\arguments{
\item{position}{A matrix or data frame of coordinates, with each row representing
a point and two columns for x and y coordinates.}

\item{mode}{A character string specifying the neighbor selection mode.
Use "dist" for distance-based filtering or "rank" for rank-based filtering.
Default is "rank".}

\item{rank.n}{An integer indicating the number of nearest neighbors to select
in rank-based mode. Default is 2.}

\item{th.min}{A numeric value specifying the minimum distance for neighbors in
distance-based mode. Spots closer than this distance will be excluded.
Default is 0.}

\item{th.max}{A numeric value specifying the maximum distance for neighbors in
distance-based mode. Spots farther than this distance will be excluded.
Default is 1.}

\item{plot}{Logical. If TRUE, plot the neighborhood of one sampled spot}
}
\value{
A list where each element contains the neighboring spots for a given spot
}
\description{
Integrate spatial neighbors based on distance or rank
}
