% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{mean_imputation}
\alias{mean_imputation}
\title{mean_imputation: mean imputation function for multi-view spectral clustering
with missing data}
\usage{
mean_imputation(l)
}
\arguments{
\item{l}{A list of data frames: all those to be included in the imputation.}
}
\value{
A list of completed data frames.
}
\description{
Works on a list of similarity matrices to impute missing values using the mean 
from the other views.
}
\examples{
m_test <- mean_imputation(misslfilled)
}
