\name{merge_predict}
\alias{merge_predict}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
predict patient group
}
\description{
predicted cell type
}
\usage{
merge_perdict(matrix, sampleID = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{matrix}{
Rows should be cells and columns should be genes.
}
  \item{sampleID}{
sampleID should be a sample list
}
}
\value{
return predicted cell type
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x)
{
  }
}
