% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{missl}
\alias{missl}
\title{A list of the blob data as similarity matrices with
a missing entry in one}
\format{A list of two data frames}
\usage{
missl
}
\description{
Two copies of a simulated dataset of 8 Gaussian blobs in a 
list converted to a similarity matrix, but one has a 
missing observation.
}
\keyword{datasets}
