\name{patient_group_predict}
\alias{patient_group_predict}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
predict patient group
}
\description{
predicted group
}
\usage{
patient_group_predict(proportion)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{proportion}{
Rows should be samples and columns should be cell types
}
}
\value{
return predicted group
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x)
{
  }
}

