% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{pca}
\alias{pca}
\title{pca: A pca function}
\usage{
pca(mydata, labels = FALSE, dotsize = 3, axistextsize = 18,
  legendtextsize = 18)
}
\arguments{
\item{mydata}{Data frame or matrix: matrix or data frame with samples as columns, features as rows}

\item{labels}{Factor: to label the plot with colours}

\item{dotsize}{Numerical value: dot size}

\item{axistextsize}{Numerical value: axis text size}

\item{legendtextsize}{Numerical value: legend text size}
}
\value{
A pca plot object
}
\description{
pca: A pca function
}
\examples{
ex_pca <- pca(blobs[,1:50])
}
