% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakAnnotation.R
\name{PeakAnnotation}
\alias{PeakAnnotation}
\alias{class:PeakAnnotation}
\alias{class:PeakAnnotations}
\alias{PeakAnnotation-class}
\alias{PeakAnnotations-class}
\alias{PeakAnnotations}
\alias{show,PeakAnnotation-method}
\alias{showAsCell,PeakAnnotations-method}
\title{Peak annotations}
\arguments{
\item{peaks}{A \code{list} of peak m/z values to be annotated.}

\item{annotations}{A \code{list} of peak annotations.}

\item{...}{Objects of class \code{PeaksAnnotation}.}
}
\description{
This class defines how peaks in an MS spectrum are annotated. In its current
preliminary implementation, if defines lists of peaks and and annotations,
both of which must have the same length. Each element of \code{peaks} is
annotated by the corresponding element of \code{annotations}. Elements of \code{peaks}
can contain a single numeric corresponding to the m/z of a peak in a
centroided spectrum, of a vector of m/z values when a set of peaks (for
instance in case of profile mode data) are annotated by an
annotation.

Currently, there are no constrains on either \code{peaks} and \code{annotations}
elements. The latters can be either descriptive characters, as in the
examples below, or could be encoded as dataframes. This is likely to change
in the future.
}
\examples{
pa1 <- PeakAnnotation(peaks = list(123.3,
                                   c(234.1, 234.12, 234.15)),
                      annotations = list("b2", "y3"))
pa1

pa2 <- PeakAnnotation(peaks = list(c(345.3, 345.32, 345.35, 345.39, 345.4),
                                   657.01,
                                   1231.182),
                      annotations = list("a5", "z3", "y7"))
pa2

pas <- PeakAnnotations(pa1, pa2)
pas

dfr <- DataFrame(x = 1:2,
                 y = LETTERS[1:2],
                 annot = pas)
dfr
}
