% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{plot_ct}
\alias{plot_ct}
\title{Visualize the spatial distribution of cell type proportion}
\usage{
plot_ct(
  proportion,
  position,
  ct.select,
  colors = c("lightblue", "lightyellow", "red"),
  NumCols = 4,
  px.size = 1,
  ratio = 1,
  max.cf = 1
)
}
\arguments{
\item{proportion}{Data frame of cell type proportion, with each row representing
a spot and each column representing a cell type}

\item{position}{A matrix or data frame of coordinates, with each row representing
a point and two columns for x and y coordinates.}

\item{ct.select}{vector containing the cell types of interest you want to plot}

\item{colors}{Vector of color names, default color scale is c("lightblue","lightyellow","red")}

\item{NumCols}{Number of columns for display when combining plots}

\item{px.size}{Adjust point size for plotting}

\item{ratio}{Numeric. Aspect ratio for the plot. Default is 1.}

\item{max.cf}{Vector of minimum and maximum cutoff values for each cell type}
}
\value{
Returns a ggplot object
}
\description{
Visualize the spatial distribution of cell type proportion
}
