% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{plot_pie}
\alias{plot_pie}
\title{Visualize the spatial distribution of cell type proportion in a geom scatterpie plot}
\usage{
plot_pie(
  proportion,
  position,
  cols,
  px.size = 0.52,
  title = NULL,
  legend = T,
  legend.title = "Cell Type",
  ratio = 1
)
}
\arguments{
\item{proportion}{Data frame of cell type proportion, with each row representing
a spot and each column representing a cell type}

\item{position}{A matrix or data frame of coordinates, with each row representing
a point and two columns for x and y coordinates.}

\item{cols}{Vector of color names}

\item{px.size}{Adjust point size for plotting}

\item{title}{Text for the title}

\item{legend}{Whether to remove the legend}

\item{legend.title}{Text for the legend title}

\item{ratio}{Numeric. Aspect ratio for the plot. Default is 1.}
}
\value{
Returns a ggplot object
}
\description{
Visualize the spatial distribution of cell type proportion in a geom scatterpie plot
}
