% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo.R
\name{pre_process}
\alias{pre_process}
\title{Pre-process gene count matrix of ST spots for downstream analysis}
\usage{
pre_process(st_data, minCountGene = 100, minCountSpot = 5, maxVol = 1)
}
\arguments{
\item{st_data}{Raw ST count data, each column is a spot and each row is a gene}

\item{minCountGene}{Minimum counts for each gene}

\item{minCountSpot}{Removes genes present in this number or more of spots}

\item{maxVol}{Removes genes present in this fraction or more of spots}
}
\value{
Returns filtered spatial count matrix
}
\description{
Pre-process gene count matrix of ST spots for downstream analysis
}
