% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{rbfkernel_b}
\alias{rbfkernel_b}
\title{rbfkernel_b: fast self-tuning kernel}
\usage{
rbfkernel_b(mat, K = 3, sigma = 1)
}
\arguments{
\item{mat}{Matrix: matrix should have samples as columns and rows as features}

\item{K}{Numerical value: the number of nearest neighbours to use when calculating local sigma}

\item{sigma}{Numerical value: a global sigma, usually left to 1 which has no effect}
}
\value{
A kernel matrix
}
\description{
rbfkernel_b: fast self-tuning kernel
}
\examples{
stsc_kern <- rbfkernel_b(blobs[,1:50])
}
