% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{sigma_finder}
\alias{sigma_finder}
\title{sigma_finder: heuristic to find sigma for the Ng kernel}
\usage{
sigma_finder(mat, NN = 3)
}
\arguments{
\item{mat}{Data frame or matrix: with points as columns, features as rows}

\item{NN}{Numerical value: the number of nearest neighbours to use (default=3)}
}
\value{
A global sigma
}
\description{
This is a heuristic to find the sigma for the kernel from the Ng spectral clustering algorithm. 
It returns a global sigma. It uses the mean K nearest neighbour distances of all samples to
determine sigma.
}
\examples{
sig <- sigma_finder(blobs)
}
