% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spirals}
\alias{spirals}
\title{Two spirals wrapped around one another}
\format{A data frame with 2 rows and 180 variables}
\usage{
spirals
}
\description{
Simulated data using the 'mlbench' CRAN package.
}
\keyword{datasets}
