% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{umap}
\alias{umap}
\title{umap: A umap function for similarity matrices or ordinary data}
\usage{
umap(mydata, labels = FALSE, dotsize = 3, similarity = TRUE,
  axistextsize = 18, legendtextsize = 18)
}
\arguments{
\item{mydata}{Data frame or matrix: kernel matrix or data frame with samples as columns, features as rows}

\item{labels}{Factor: to label the plot with colours}

\item{dotsize}{Numerical value: dot size}

\item{similarity}{Logical flag: whether input is similarity matrix or not}

\item{axistextsize}{Numerical value: axis text size}

\item{legendtextsize}{Numerical value: legend text size}
}
\value{
A umap plot object
}
\description{
umap: A umap function for similarity matrices or ordinary data
}
\examples{
ex_umap <- umap(blobs[,1:50],similarity=FALSE)
}
