% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brain}
\alias{brain}
\title{A brain cancer dataset}
\format{A list of data frames}
\source{
\url{https://gdac.broadinstitute.org/}
}
\usage{
brain
}
\description{
A dataset containing The Cancer Genome Atlas expression
data. From this publication https://tcga-data.nci.nih.gov/docs/publications/lgggbm_2016/.
The first data frame is a 5133X150 RNA-seq data matrix, the second is a 262X150
miRNA-seq data matrix, the third is 45X150 protein array data matrix. The data was
all pre-normalised then subject to log transform.
}
\keyword{datasets}
