% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{community_detect}
\alias{community_detect}
\title{Detect spatial communities based on proportion matrix and spatial positions}
\usage{
community_detect(
  pred_prop,
  position,
  shape = "polygon",
  mode = "dist",
  max_nbh = 3,
  rank.n = 3,
  threshList = c(0, 140, 280),
  detect_res = 0.4,
  refine = F
)
}
\arguments{
\item{pred_prop}{A matrix containing cell type proportions, with rows representing
spots and columns representing cell types.}

\item{position}{A matrix or data frame of coordinates, with each row representing
a point and two columns for x and y coordinates.}

\item{shape}{A character string specifying the shape of the spatial distribution
("regular polygon" or "random").
The number of neighbors to consider is set based on this shape.
Default is "polygon"}

\item{mode}{A character string specifying the neighbor selection mode.
Use "dist" for distance-based filtering or "rank" for rank-based filtering.
Default is "rank".}

\item{max_nbh}{An integer indicating the maximum number of neighbors for rank-based
mode (default is 3).}

\item{rank.n}{An integer indicating the number of nearest neighbors to select
in rank-based mode. Default is 2.}

\item{threshList}{A numeric value specifying the minimum distance for neighbors in
distance-based mode. Spots closer than this distance will be excluded.
Default is 0.}

\item{detect_res}{A numeric value specifying the resolution parameter for clustering (default is 0.4).}

\item{refine}{whether to refine the results (default is False).}
}
\value{
A character vector containing refined community labels for each spatial spot.
}
\description{
Detect spatial communities based on proportion matrix and spatial positions
}
