% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo.R
\name{deconvo_process}
\alias{deconvo_process}
\title{Reference-free deconvolution for spatial transcriptomics by SPECTRUM}
\usage{
deconvo_process(
  st_data,
  location,
  maxVol = 0.975,
  data_struct = "dist",
  dist.th = 2,
  deldir.nb = 3,
  markerList
)
}
\arguments{
\item{st_data}{A matrix or data frame containing spatial transcriptomics data
with genes as rows and spatial spots as columns.}

\item{location}{A matrix or data frame of coordinates, with each row representing
a point and two columns for x and y coordinates.}

\item{maxVol}{Removes genes present in this fraction or more of spots}

\item{data_struct}{A character string specifying the structure to use.
Options are "deldir" for Delaunay triangulation
or "dist" (default) for distance threshold.}

\item{dist.th}{A numeric threshold for the distance used when data_struct is set
to "dist". Defaults to 2.}

\item{deldir.nb}{An integer specifying the threshold for the number of neighbors to
consider in the Gabriel graph construction. Defaults to 3.}

\item{markerList}{A list where each element is a vector of markers
for a specific cell type.}
}
\value{
A matrix representing the proportions of each cell type in each spatial
spot. Rows represent spatial spots and columns represent cell types.
}
\description{
Reference-free deconvolution for spatial transcriptomics by SPECTRUM
}
