% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvo.R
\name{get_localization_score}
\alias{get_localization_score}
\title{Calculate Localization Score for Genes Based on ST data}
\usage{
get_localization_score(
  st_data,
  location,
  maxVol,
  data_struct = "dist",
  dist.th = 2,
  deldir.nb = 3
)
}
\arguments{
\item{st_data}{A matrix or data frame containing spatial transcriptomics data
with genes as rows and spatial spots as columns.}

\item{location}{A matrix or data frame of coordinates, with each row representing
a point and two columns for x and y coordinates.}

\item{maxVol}{Removes genes present in this fraction or more of spots. Default is 0.975.}

\item{data_struct}{A character string specifying the structure to use.
Options are "deldir" for Delaunay triangulation
or "dist" (default) for distance threshold.}

\item{dist.th}{A numeric threshold for the distance used when data_struct is set
to "dist". Defaults to 2.}

\item{deldir.nb}{An integer specifying the threshold for the number of neighbors to
consider in the Gabriel graph construction. Defaults to 3.}
}
\value{
A named numeric vector where each element is the calculated localization
score for a gene.
}
\description{
Calculate Localization Score for Genes Based on ST data
}
