% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{kernel_pca}
\alias{kernel_pca}
\title{kernel_pca: A kernel pca function}
\usage{
kernel_pca(datam, labels = FALSE, axistextsize = 18,
  legendtextsize = 18, dotsize = 3, similarity = TRUE)
}
\arguments{
\item{datam}{Dataframe or matrix: a data frame with samples as columns, rows as features, or a kernel matrix}

\item{labels}{Factor: to label the plot with colours}

\item{axistextsize}{Numerical value: axis text size}

\item{legendtextsize}{Numerical value: legend text size}

\item{dotsize}{Numerical value: dot size}

\item{similarity}{Logical flag: whether the input is a similarity matrix or not}
}
\value{
A kernel PCA plot
}
\description{
kernel_pca: A kernel pca function
}
\examples{
ex_kernel_pca <- kernel_pca(blobs[,1:50], similarity=FALSE)
}
