% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{refine_res}
\alias{refine_res}
\title{Refine predicted communities based on spatial neighbors}
\usage{
refine_res(spots, pred_res, position, neighbors)
}
\arguments{
\item{spots}{A character vector of spot names.}

\item{pred_res}{A character or factor vector of initial predicted identities
for each spot, with length equal to the length of spots.}

\item{position}{A data frame or matrix containing spatial coordinates, with rows
corresponding to spots and columns representing x and y coordinates.}

\item{neighbors}{List of spot neighbors calculated in the integrate step.}
}
\value{
A character vector of refined predictions for each spot
}
\description{
Refine predicted communities based on spatial neighbors
}
